Revert "Changed type of strikeprice from double to decimal to be consistent with the recent changes in tbloptions in TMLDB."
This reverts commit ab53b960db7bf2b7e77dcc2abf0d9dc825bd9d4d.

+++ ICE_Import/ICE_Import/PushTable.cs
@@ -155,7 +155,7 @@ namespace ICE_Import
             tblOptions.Columns.Add("optionmonth", typeof(char));
             tblOptions.Columns.Add("optionmonthint", typeof(int));
             tblOptions.Columns.Add("optionyear", typeof(int));
-            tblOptions.Columns.Add("strikeprice", typeof(decimal));
+            tblOptions.Columns.Add("strikeprice", typeof(float));
             tblOptions.Columns.Add("callorput", typeof(char));
             tblOptions.Columns.Add("idinstrument", typeof(int));
             tblOptions.Columns.Add("expirationdate", typeof(DateTime));
@@ -185,7 +185,7 @@ namespace ICE_Import
                         CqgSymbol,
                         monthChar,
                         option.StripName.Year,
-                        (double)option.StrikePrice.GetValueOrDefault(),
+                        option.StrikePrice.GetValueOrDefault(),
                         0,
                         0,
                         IdInstrument);
